+++ NEWS.md
@@ -1,6 +1,6 @@
 # bbotk 0.5.1.9000
 
-* feat: Added `$label` to `Optimizer`s and `Terminators`.
+* feat: Added `$label` to `Optimizer`s and `Terminator`s.
 
 # bbotk 0.5.1